Add files via upload
This Matlab code extracts timestamps from a log file and calculates the duration of time covered by those timestamps. The log file can contain arbitrary text, but the code expects timestamps to be in the format of "mm/dd/yyyy - hh:mm:ss AM/PM". The code outputs the start and end times, and the duration of time covered by the timestamps in seconds.